allow resourceType and filter to both be specified
One quirk about the ARM resource list API is that the $filter
specification requires the resourceType be used in very specific
ways as hilighted in this issue:

  https://github.com/Azure/azure-rest-api-specs/issues/21324

Due to that quick, azure-metrics-exporter disallowed scrape
configs specifying both resourceType and filter.

It would be useful to allow specifying both resourceType and
filter for cases where the user wish to target only a subset
of resources of a given type. The following example was a case
where the user wanted to do that, but ended up working around
the issue by using the substringof $filter keyword.

  https://github.com/webdevops/azure-metrics-exporter/issues/51

The substringof $filter keyword only works for targetting
resource name and resource group, and it does not work for
targetting resource tag names or values.

This update makes speciying both resourceType and filter to become
allowed by filtering for the specified resourceType locally.